remove headless test ignore, move test
