+++ tox.ini
@@ -81,7 +81,9 @@ commands_pre =
     headless: pip uninstall -y pytest-qt qtpy pyqt5 pyside2
 commands =
     !headless: python -m pytest {env:PYTEST_PATH:} --color=yes --basetemp={envtmpdir} --cov-report=xml --cov={env:PYTEST_PATH:napari} --ignore tools {posargs}
-    headless: python -m pytest --color=yes --basetemp={envtmpdir} --ignore napari/_vispy --ignore napari/_qt --ignore napari/_tests --ignore tools --ignore napari/layers/utils/_tests/test_interactivity_utils.py {posargs}
+    # do not add ignores to this line just to make headless tests pass.
+    # nothing outside of _qt or _vispy should require Qt or make_napari_viewer
+    headless: python -m pytest --color=yes --basetemp={envtmpdir} --ignore napari/_vispy --ignore napari/_qt --ignore napari/_tests --ignore tools {posargs}
 
 
 [testenv:isort]
